Updated KeyDataPairsToDictionaryStringObjectJsonConverter to supprrt when data is null, and added some tests
Updated audit log to use a double when ValueKind is a number to support real numbers, and added some tests

Updated document to documents

Fixed an issue with concurrency within Linking due to the etag being empty

+++ Cdms.Model/Auditing/AuditEntry.cs
@@ -149,7 +149,7 @@ public class AuditEntry
                         value = operation.Value.GetValue<string>();
                         break;
                     case JsonValueKind.Number:
-                        value = operation.Value.GetValue<long>();
+                        value = operation.Value.GetValue<double>();
                         break;
                     case JsonValueKind.True:
                     case JsonValueKind.False:
